fixed Settings and Onbaording Bugs, Changed Project Structure to MVVM
