+++ macOS/writing-tools/App/AppDelegate.swift
@@ -54,7 +54,7 @@ class AppDelegate: NSObject, NSApplicationDelegate, NSWindowDelegate {
                 self?.recreateStatusBarItem()
             }
             
-            if !UserDefaults.standard.bool(forKey: "has_completed_onboarding") {
+            if UserDefaults.standard.bool(forKey: "has_completed_onboarding") {
                 self?.showOnboarding()
             }
