+++ macOS/writing-tools/App/AppSettings.swift
