+++ macOS/writing-tools/App/AppState.swift
