+++ macOS/writing-tools/Models/AI Providers/GeminiProvider.swift
