+++ macOS/writing-tools/Models/AI Providers/OllamaProvider.swift
