+++ macOS/writing-tools/Models/CommandModel.swift
