+++ macOS/writing-tools/Models/CustomCommand.swift
