+++ macOS/writing-tools/Models/WritingOption.swift
