+++ macOS/writing-tools/Services/CommandManager.swift
