+++ macOS/writing-tools/Services/UpdateChecker.swift
