+++ macOS/writing-tools/Services/WindowManager.swift
