+++ macOS/writing-tools/Views/Chat/ResponseView.swift
