+++ macOS/writing-tools/Views/Chat/ResponseWindow.swift
