+++ macOS/writing-tools/Views/ContentView.swift
